requirements.txt
Django==1.5
Markdown==2.2.1
argparse==1.2.1
distribute==0.6.24
django-filter==0.5.4
djangorestframework==2.2.3
wsgiref==0.1.2
yolk==0.4.3
